Move headless specific stuff to a separate namespace
