+++ src/Microsoft.DotNet.XHarness.CLI/CommandArguments/AndroidHeadless/AndroidHeadlessInstallCommandArguments.cs
@@ -4,10 +4,11 @@
 
 using System;
 using System.Collections.Generic;
+using Microsoft.DotNet.XHarness.CLI.CommandArguments.Android;
 
-namespace Microsoft.DotNet.XHarness.CLI.CommandArguments.Android;
+namespace Microsoft.DotNet.XHarness.CLI.CommandArguments.AndroidHeadless;
 
-internal class AndroidHeadlessRunCommandArguments : XHarnessCommandArguments, IAndroidHeadlessAppRunArguments
+internal class AndroidHeadlessInstallCommandArguments : XHarnessCommandArguments, IAndroidHeadlessAppRunArguments
 {
     public TestPathArgument TestPath { get; } = new();
     public RuntimePathArgument RuntimePath { get; } = new();
@@ -17,10 +18,8 @@ internal class AndroidHeadlessRunCommandArguments : XHarnessCommandArguments, IA
     public TimeoutArgument Timeout { get; } = new(TimeSpan.FromMinutes(15));
     public LaunchTimeoutArgument LaunchTimeout { get; } = new(TimeSpan.FromMinutes(5));
     public DeviceIdArgument DeviceId { get; } = new();
+    public DeviceArchitectureArgument DeviceArchitecture { get; } = new();
     public ApiVersionArgument ApiVersion { get; } = new();
-    public ExpectedExitCodeArgument ExpectedExitCode { get; } = new((int)Common.CLI.ExitCode.SUCCESS);
-    public DeviceOutputFolderArgument DeviceOutputFolder { get; } = new();
-    public WifiArgument Wifi { get; } = new();
 
     protected override IEnumerable<Argument> GetArguments() => new Argument[]
     {
@@ -30,11 +29,9 @@ internal class AndroidHeadlessRunCommandArguments : XHarnessCommandArguments, IA
         TestScript,
         OutputDirectory,
         Timeout,
-        LaunchTimeout,
         DeviceId,
+        LaunchTimeout,
+        DeviceArchitecture,
         ApiVersion,
-        ExpectedExitCode,
-        DeviceOutputFolder,
-        Wifi,
     };
 }
