+++ src/Microsoft.DotNet.XHarness.CLI/CommandArguments/AndroidHeadless/Arguments/RuntimePathArgument.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.DotNet.XHarness.CLI.CommandArguments.Android;
+namespace Microsoft.DotNet.XHarness.CLI.CommandArguments.AndroidHeadless;
 
 internal class RuntimePathArgument : RequiredStringArgument
 {
