+++ src/Microsoft.DotNet.XHarness.CLI/CommandArguments/AndroidHeadless/Arguments/TestAssemblyArgument.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.DotNet.XHarness.CLI.CommandArguments.Android;
+namespace Microsoft.DotNet.XHarness.CLI.CommandArguments.AndroidHeadless;
 
 internal class TestAssemblyArgument : RequiredStringArgument
 {
