+++ src/Microsoft.DotNet.XHarness.CLI/CommandArguments/AndroidHeadless/IAndroidHeadlessAppRunArguments.cs
@@ -1,8 +1,9 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
+using Microsoft.DotNet.XHarness.CLI.CommandArguments.Android;
 
-namespace Microsoft.DotNet.XHarness.CLI.CommandArguments.Android;
+namespace Microsoft.DotNet.XHarness.CLI.CommandArguments.AndroidHeadless;
 
 internal interface IAndroidHeadlessAppRunArguments
 {
