+++ src/Microsoft.DotNet.XHarness.CLI/Commands/AndroidHeadless/AndroidHeadlessInstallCommand.cs
@@ -4,12 +4,14 @@ using System.IO;
 using System.Linq;
 using Microsoft.DotNet.XHarness.Android;
 using Microsoft.DotNet.XHarness.CLI.Android;
+using Microsoft.DotNet.XHarness.CLI.AndroidHeadless;
 using Microsoft.DotNet.XHarness.CLI.CommandArguments.Android;
+using Microsoft.DotNet.XHarness.CLI.CommandArguments.AndroidHeadless;
 using Microsoft.DotNet.XHarness.Common;
 using Microsoft.DotNet.XHarness.Common.CLI;
 using Microsoft.Extensions.Logging;
 
-namespace Microsoft.DotNet.XHarness.CLI.Commands.Android;
+namespace Microsoft.DotNet.XHarness.CLI.Commands.AndroidHeadless;
 
 internal class AndroidHeadlessInstallCommand : AndroidCommand<AndroidHeadlessInstallCommandArguments>
 {
