+++ src/Microsoft.DotNet.XHarness.CLI/Commands/AndroidHeadless/AndroidHeadlessUninstallCommand.cs
@@ -4,11 +4,13 @@
 
 using Microsoft.DotNet.XHarness.Android;
 using Microsoft.DotNet.XHarness.CLI.Android;
+using Microsoft.DotNet.XHarness.CLI.AndroidHeadless;
 using Microsoft.DotNet.XHarness.CLI.CommandArguments.Android;
+using Microsoft.DotNet.XHarness.CLI.CommandArguments.AndroidHeadless;
 using Microsoft.DotNet.XHarness.Common.CLI;
 using Microsoft.Extensions.Logging;
 
-namespace Microsoft.DotNet.XHarness.CLI.Commands.Android;
+namespace Microsoft.DotNet.XHarness.CLI.Commands.AndroidHeadless;
 
 internal class AndroidHeadlessUninstallCommand : AndroidCommand<AndroidHeadlessUninstallCommandArguments>
 {
